requirements.txt
chromadb
litellm
python-dotenv
pydantic
sentence-transformers
tenacity
tqdm
